+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # unified-latex Changelog
 
+### v1.5.0
+- HTML conversion: `vspace` and `hspace` now give the amount in a `data-amount` attribute.
+- HTML conversion: unknown macros now have their arguments wrapped in spans instead of appearing as formatted LaTeX code.
+- Add basic Markdown conversion support.
+
+### v1.4.2
+- Avoid slowdown when paring incomplete environments (e.g. `\newcommand{\x}{\begin{x}}`). This is accomplished by enabling caching in PEGjs.
+- Added `"` ligature and `\paragraph` and `\subparagraph` to HTML conversion.
+
 ### v1.4.1
 - Many more ligatures added to the HTML converter.
 - Fixed issue [#40](https://github.com/siefkenj/unified-latex/issues/40) where the optional argument to `\\` was being parsed even if preceded by a space. (E.g., `\\[10pt]` and `\\ [10pt]`) were parsed the same. Not allowing the space should more closely match expected behavior.